Wire Nyx.Combat into live combat and remove Nyx.AttackEngine
Replaces the Nyx.AttackEngine integration with Nyx.Combat and deletes the
old module (23 files) along with its solution entries and project reference.

Host-side seams, all new:
  * CombatProjection  - Entity -> Combatant snapshot; the only file that
    knows both models. Handles the enum collision between EntityFlag.Bot=3
    and EntityKind.Structure=3, inverts ItemBless (a multiplier) into the
    engine's percentage reduction, and packs Gems[] into PackedGems.
  * NyxCombatWorld    - ICombatWorld over Kernel.Maps + Kernel.GamePool,
    with hostility delegated to Handle.CanAttack so the engine can never
    disagree with the rest of the server about who may hit whom. Guards the
    unconditional attacker.Owner deref in that method for monster attackers.
  * CombatAdapter     - MsgAttack <-> engine translation. Declines rather
    than rejects anything it does not fully own, so the legacy Handle path
    still runs unchanged; resolved damage is handed to Handle.ReceiveAttack
    so experience/durability/death/PK stay in one place.
  * AddCombat()       - composition root replacing AddAttackEngine().

Fixes carried along the way:
  * PacketHandler 1022: the engine hook's unbraced 'break' after Log.Warning
    fired whenever GamePool contained the target, skipping Handle even on a
    successful resolve. The hook also resolved two services from the DI
    container per attack packet; both are now cached.
  * Combatant refinery stats widened byte -> ushort. The live server stores
    them in hundredths of a percent (a refined character carries CriticalStrike
    1500), which a byte silently truncated to 220.
  * Refinery crit/penetration/skill-crit rolls were off by 100x against
    Calculate.Refinary and would have proc'd on essentially every hit.

Config: Combat.UseAttackEngine -> Combat.UseCombatEngine, still default false.
The catalog loads from cq_magictype at boot either way, so a bad table is
caught at startup rather than on the first attack.

No .NET SDK is available in this environment, so this is inspection-verified
rather than compiled.

+++ Modules/Nyx.Combat/Model/Combatant.cs
@@ -101,15 +101,22 @@ public readonly struct Combatant
     /// <summary>Armour weight; feeds the ranged evasion term.</summary>
     public int Weight { get; init; }
 
-    // ---- refinery / Chi attributes, all in percent ----------------------
+    // ---- refinery / Chi attributes --------------------------------------
 
-    public byte CriticalStrike { get; init; }
-    public byte SkillCriticalStrike { get; init; }
-    public byte Immunity { get; init; }
-    public byte Penetration { get; init; }
-    public byte Breakthrough { get; init; }
-    public byte Counteraction { get; init; }
-    public byte ShieldBlockPercent { get; init; }
+    /// <remarks>
+    /// These are <see cref="ushort"/> rather than <see cref="byte"/> because the
+    /// live server stores them in hundredths of a percent — a fully refined
+    /// character carries a Critical Strike of 1500, which a byte would silently
+    /// truncate to 220. The divisors in
+    /// <see cref="Configuration.CombatOptions"/> convert them to a roll.
+    /// </remarks>
+    public ushort CriticalStrike { get; init; }
+    public ushort SkillCriticalStrike { get; init; }
+    public ushort Immunity { get; init; }
+    public ushort Penetration { get; init; }
+    public ushort Breakthrough { get; init; }
+    public ushort Counteraction { get; init; }
+    public ushort ShieldBlockPercent { get; init; }
 
     // ---- damage percentages --------------------------------------------
